+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Versioning](https://semver.org/spec/v2.0.0.html) for `libnopegl`.
 - `Texture.premult` parameter to premultiply ngl_texvideo() output color by its
   alpha
 - `ngl_error_to_string()` to convert error codes to string messages
+- `ngl_node_param_swap_elem` to swap elements within a list
 
 ### Fixed
 - Crash when using resizable RTTs with time ranges